+++ packages/terra-tabs/src/common-tabs/_Tabs.jsx
@@ -1,11 +1,15 @@
+/* eslint-disable react/no-access-state-in-setstate */
 /* eslint-disable class-methods-use-this */
 import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
 import ThemeContext from 'terra-theme-context';
 import ResizeObserver from 'resize-observer-polyfill';
+import { v4 as uuid } from 'uuid';
+import IconAdd from 'terra-icon/lib/icon/IconAdd';
 import { DragDropContext, Droppable } from 'react-beautiful-dnd';
 import { injectIntl } from 'react-intl';
+import AddButton from './_AddButton';
 import MoreButton from './_MoreButton';
 import TabDropDown from './_TabDropDown';
 import Tab from './_Tab';
@@ -19,6 +23,10 @@ const propTypes = {
    * The label to set on the tablist element.
    */
   ariaLabel: PropTypes.string.isRequired,
+  /**
+   * The label to set on the add icon element.
+   */
+  ariaLabelAddTab: PropTypes.string,
   /**
    * Currently active Tabs.Pane content to be displayed.
    */
@@ -60,12 +68,33 @@ const propTypes = {
    */
   onChange: PropTypes.func,
   /**
-   * Whether or not the tab is draggable.
+   * Callback function when add button selection has changed.
+   * Parameters: 1. Event 2. Selected pane's key
    */
+  onSelectAddButton: PropTypes.func,
+  /**
+   * Callback function triggered when close button clicked.
+   * It receives two parameters: 1. Item key of the tab to close. 2. MetaData associated with the tab.
+   */
+  onClosingTab: PropTypes.func,
+
+  /**
+   * Indicates whether the tabs are closable, allowing the user to close tabs.
+   */
+  isClosable: PropTypes.bool,
+
+  /**
+   * Callback function to handle changes in the tab state.
+   * It receives the updated tab data after a tab is closed.
+   */
+  onTabStateChange: PropTypes.func,
+  /**
+  * Whether or not the tab is draggable.
+  */
   isDraggable: PropTypes.bool,
   /**
   * Callback function triggered when tab is drag and dropped .
-   */
+  */
   onTabOrderChange: PropTypes.func,
   /**
    * @private
@@ -80,7 +109,7 @@ class Tabs extends React.Component {
     this.containerRef = React.createRef();
     this.dropdownRef = React.createRef();
     this.moreButtonRef = React.createRef();
-
+    this.addButtonRef = React.createRef();
     this.setIsOpen = this.setIsOpen.bind(this);
     this.resetCache = this.resetCache.bind(this);
     this.handleResize = this.handleResize.bind(this);
@@ -91,6 +120,8 @@ class Tabs extends React.Component {
     this.handleOutsideClick = this.handleOutsideClick.bind(this);
     this.wrapOnSelect = this.wrapOnSelect.bind(this);
     this.wrapOnSelectHidden = this.wrapOnSelectHidden.bind(this);
+    this.wrapOnSelectHidden = this.wrapOnSelectHidden.bind(this);
+    this.wrapOnAddButton = this.wrapOnAddButton.bind(this);
     this.positionDropDown = this.positionDropDown.bind(this);
     this.handleDragEnd = this.handleDragEnd.bind(this);
     this.handleDragStart = this.handleDragStart.bind(this);
@@ -101,6 +132,7 @@ class Tabs extends React.Component {
     this.state = {
       visibleTabData: this.props.tabData,
     };
+    this.isNewTabCreated = false;
   }
 
   componentDidMount() {
@@ -114,19 +146,41 @@ class Tabs extends React.Component {
       }
     });
     this.resizeObserver.observe(this.containerRef.current);
+    this.resizeObserver.observe(this.containerRef.current.parentNode);
     this.handleResize();
+    this.addTabId = uuid();
   }
 
   componentDidUpdate(prevProps) {
     const prevTab = prevProps.tabData.find((tab) => tab.isSelected === true);
     const currTab = this.props.tabData.find((tab) => tab.isSelected === true);
+    const prevtabKeys = [];
+    prevProps.tabData.forEach(child => {
+      prevtabKeys.push(child.id);
+    });
+    const curtabKeys = [];
+    this.props.tabData.forEach(child => {
+      curtabKeys.push(child.id);
+    });
 
     // Allow dynamic addition of tabs.
     if (this.state.visibleTabData.length !== this.props.tabData.length) {
       // eslint-disable-next-line react/no-did-update-set-state
       this.setState({ visibleTabData: this.props.tabData });
+    } else {
+      let isTabEqual = false;
+      for (let i = 0; i < curtabKeys.length; i += 1) {
+        const prevKey = prevtabKeys[i];
+        const curKey = curtabKeys[i];
+        if (prevKey !== curKey) {
+          isTabEqual = true;
+        }
+      }
+      if (isTabEqual) {
+        // eslint-disable-next-line react/no-did-update-set-state
+        this.setState({ visibleTabData: this.props.tabData });
+      }
     }
-
     // Allow Active Styles to be applied when tab is selected.
     if (prevTab && currTab && prevTab.id !== currTab.id) {
       // eslint-disable-next-line react/no-did-update-set-state
@@ -143,6 +197,19 @@ class Tabs extends React.Component {
         return { visibleTabData: newArray };
       });
     }
+
+    if (currTab && !this.moreButtonRef.current) {
+      const element = document.getElementById(currTab.id);
+      if (element && this.isNewTabCreated) {
+        element.focus();
+      }
+    } else if (this.state.visibleTabData.length === 0) {
+      const element = document.getElementById(this.addTabId);
+      if (element) {
+        element.focus();
+      }
+    }
+
     if (this.isCalculating) {
       this.isCalculating = false;
       this.handleResize(this.contentWidth);
@@ -172,7 +239,7 @@ class Tabs extends React.Component {
     const availableWidth = width - moreButtonWidth;
 
     // Calculate hidden index
-    const tabCount = this.props.tabData.length;
+    const tabCount = this.state.visibleTabData.length;
     let newHideIndex = tabCount;
     let calcMinWidth = 0;
     let showMoreButton = false;
@@ -225,7 +292,7 @@ class Tabs extends React.Component {
   }
 
   handleOutsideClick(event) {
-    if (event.type === 'mousedown' && (this.moreButtonRef.current === event.currentTarget || this.moreButtonRef.current.contains(event.currentTarget))) {
+    if (this.moreButtonRef.current && event.type === 'mousedown' && (this.moreButtonRef.current === event.currentTarget || this.moreButtonRef.current.contains(event.currentTarget))) {
       return;
     }
     this.setIsOpen(false);
@@ -272,6 +339,7 @@ class Tabs extends React.Component {
     if (!result.destination) {
       return;
     }
+    this.isNewTabCreated = false;
     this.setState((prevState) => {
       const items = this.reorder(
         prevState.visibleTabData,
@@ -294,7 +362,6 @@ class Tabs extends React.Component {
     const result = Array.from(list);
     const [removed] = result.splice(startIndex, 1);
     result.splice(endIndex, 0, removed);
-
     return result;
   };
 
@@ -336,23 +403,74 @@ class Tabs extends React.Component {
 
   wrapOnSelect(onSelect) {
     return (itemKey, metaData) => {
-      this.setIsOpen(false);
-      onSelect(itemKey, metaData);
+      const updatedTabData = this.state.visibleTabData.map((tab) => ({
+        ...tab,
+        isSelected: tab.itemKey === itemKey,
+      }));
+      this.setState({ visibleTabData: updatedTabData }, () => {
+        this.setIsOpen(false);
+        onSelect(itemKey, metaData);
+      });
     };
   }
 
   wrapOnSelectHidden(onSelect) {
     return (itemKey, metaData) => {
       if (this.isOpen) {
-        onSelect(itemKey, metaData);
+        const updatedTabData = this.state.visibleTabData.map((tab) => ({
+          ...tab,
+          isSelected: tab.id === itemKey,
+        }));
+        this.setState({ visibleTabData: updatedTabData }, () => {
+          onSelect(itemKey, metaData);
+          this.setIsOpen(!this.isOpen);
+        });
       }
-      this.setIsOpen(!this.isOpen);
     };
+  }
+
+  wrapOnClose() {
+    return (itemKey, event) => {
+      this.isNewTabCreated = false;
+      this.setIsOpen(false);
+      let removedTabIndex = -1;
+      const updatedTabData = this.state.visibleTabData
+        .map((tab, index) => {
+          if (tab.itemKey === itemKey && tab.isSelected === true) {
+            removedTabIndex = index;
+          }
+          return { ...tab, originalIndex: index };
+        })
+        .filter((tab) => tab.itemKey !== itemKey);
+
+      if (!updatedTabData.some((tab) => tab.isSelected === true) && updatedTabData.length > 0 && removedTabIndex !== 0 && updatedTabData[removedTabIndex - 1] !== undefined && updatedTabData[removedTabIndex - 1].isDisabled !== true) {
+        if (updatedTabData[removedTabIndex - 1].isSelected !== undefined) {
+          updatedTabData[removedTabIndex - 1].isSelected = true;
+        }
+      } else if (!updatedTabData.some((tab) => tab.isSelected === true) && updatedTabData.length > 0 && removedTabIndex !== 0 && updatedTabData[removedTabIndex - 1] !== undefined && updatedTabData[removedTabIndex - 1].isDisabled === true) {
+        if (updatedTabData[removedTabIndex - 2]?.isSelected !== undefined) {
+          updatedTabData[removedTabIndex - 2].isSelected = true;
+        }
+      } else if (removedTabIndex >= 0 && removedTabIndex < updatedTabData.length) {
+        if (updatedTabData[removedTabIndex]?.isSelected !== undefined) {
+          updatedTabData[removedTabIndex].isSelected = true;
+        }
+      }
+      this.props.onTabStateChange(updatedTabData, itemKey, event);
+      this.setState({ visibleTabData: updatedTabData });
+    };
+  }
+
+  wrapOnAddButton() {
+    if (this.props.onSelectAddButton) {
+      this.isNewTabCreated = true;
+      this.props.onSelectAddButton();
+    }
   }
 
   render() {
     const {
-      ariaLabel, variant, onChange, isDraggable,
+      ariaLabel, variant, onChange, onSelectAddButton, ariaLabelAddTab, isDraggable,
     } = this.props;
     const theme = this.context;
     const enabledTabs = this.state.visibleTabData.filter(tab => !tab.isDisabled);
@@ -360,6 +478,8 @@ class Tabs extends React.Component {
     const hiddenIds = [];
     const visibleTabs = [];
     const hiddenTabs = [];
+    const moreIds = ids;
+    if (onSelectAddButton) { moreIds.push(this.addTabId); }
     let isHiddenSelected = false;
 
     let enabledTabsIndex = -1;
@@ -384,6 +504,7 @@ class Tabs extends React.Component {
             setDropdownOpen={this.handleHiddenFocus}
             hiddenStartIndex={this.hiddenStartIndex}
             showIcon={tab.showIcon}
+            onClosingTab={this.wrapOnClose()}
             isDraggable={isDraggable}
           />,
         );
@@ -400,6 +521,7 @@ class Tabs extends React.Component {
             onChange={onChange}
             icon={tab.icon}
             showIcon={tab.showIcon}
+            onClosingTab={this.wrapOnClose()}
             isDisabled={tab.isDisabled}
             isDraggable={isDraggable}
           />,
@@ -408,6 +530,25 @@ class Tabs extends React.Component {
 
         if (tab.isSelected) {
           isHiddenSelected = true;
+        }
+        if (index === this.state.visibleTabData.length - 1 && onSelectAddButton) {
+          hiddenTabs.push(
+            <HiddenTab
+              id={this.addTabId}
+              data-focus-styles-enabled
+              itemKey={this.addTabId}
+              label={ariaLabelAddTab}
+              index={enabledTabsIndex + 1}
+              showIcon
+              icon={<IconAdd />}
+              tabIds={moreIds}
+              onSelect={this.wrapOnAddButton}
+              onFocus={this.handleHiddenFocus}
+              onBlur={this.handleHiddenBlur}
+              onChange={onChange}
+              showAddButton
+            />,
+          );
         }
       }
     });
@@ -423,6 +564,9 @@ class Tabs extends React.Component {
       };
     }
     const commonTabsClassNames = cx('tab-container', theme.className);
+    const commonTabsContainerClassNames = cx('container', theme.className);
+    const commonDivClassNames = cx('divcontainer', theme.className);
+
     window['__react-beautiful-dnd-disable-dev-warnings'] = true;
 
     if (isDraggable) {
@@ -430,46 +574,61 @@ class Tabs extends React.Component {
         <DragDropContext onDragStart={this.handleDragStart} onDragUpdate={this.handleDragUpdate} onDragEnd={this.handleDragEnd}>
           <Droppable className={commonTabsClassNames} droppableId="tab-list" direction="horizontal">
             {(provided) => (
-              <div
-                {...attrs}
-                {...provided.droppableProps}
-                ref={(el) => {
-                  provided.innerRef(el);
-                  this.containerRef.current = el; // Store the reference to the container element
-                }}
-                className={commonTabsClassNames}
-                role="tablist"
-                aria-label={ariaLabel}
-                aria-orientation="horizontal"
-                aria-owns={hiddenIds.join(' ')}
-                data-terra-drag-focus
-              >
-                {visibleTabs}
-                {provided.placeholder}
-                {this.showMoreButton ? (
-                  <MoreButton
-                    isOpen={this.isOpen}
-                    hiddenIndex={this.hiddenStartIndex}
-                    isActive={isHiddenSelected}
-                    zIndex={this.state.visibleTabData.length - this.hiddenStartIndex}
-                    onBlur={this.handleMoreButtonBlur}
-                    onSelect={this.handleMoreButtonSelect}
-                    refCallback={node => { this.moreButtonRef.current = node; }}
-                    tabIds={ids}
-                    variant={variant}
-                  />
-                ) : undefined}
-                <TabDropDown
-                  onFocus={this.handleHiddenFocus}
-                  onBlur={this.handleHiddenBlur}
-                  isOpen={this.isOpen}
-                  onRequestClose={this.handleOutsideClick}
-                  refCallback={node => { this.dropdownRef.current = node; }}
-                  isDraggable={isDraggable}
-                  onTabOrderChange={this.handleTabOrderChange}
+              <div className={commonTabsContainerClassNames}>
+                <div
+                  {...attrs}
+                  {...provided.droppableProps}
+                  ref={(el) => {
+                    provided.innerRef(el);
+                    this.containerRef.current = el; // Store the reference to the container element
+                  }}
+                  className={commonTabsClassNames}
+                  role="tablist"
+                  aria-label={ariaLabel}
+                  aria-orientation="horizontal"
+                  aria-owns={hiddenIds.join(' ')}
+                  data-terra-drag-focus
                 >
-                  {hiddenTabs}
-                </TabDropDown>
+                  {visibleTabs}
+                  {provided.placeholder}
+                  {this.showMoreButton ? (
+                    <MoreButton
+                      isOpen={this.isOpen}
+                      hiddenIndex={this.hiddenStartIndex}
+                      isActive={isHiddenSelected}
+                      zIndex={this.state.visibleTabData.length - this.hiddenStartIndex}
+                      onBlur={this.handleMoreButtonBlur}
+                      onSelect={this.handleMoreButtonSelect}
+                      refCallback={node => { this.moreButtonRef.current = node; }}
+                      tabIds={onSelectAddButton ? moreIds : ids}
+                      variant={variant}
+                    />
+                  ) : undefined}
+                  <TabDropDown
+                    onFocus={this.handleHiddenFocus}
+                    onBlur={this.handleHiddenBlur}
+                    isOpen={this.isOpen}
+                    onRequestClose={this.handleOutsideClick}
+                    refCallback={node => { this.dropdownRef.current = node; }}
+                    isDraggable={isDraggable}
+                    onTabOrderChange={this.handleTabOrderChange}
+                  >
+                    {hiddenTabs}
+                  </TabDropDown>
+                </div>
+                <div className={commonDivClassNames} ref={this.addButtonRef}>
+                  {(!this.showMoreButton && onSelectAddButton) && (
+                  <AddButton
+                    id={this.addTabId}
+                    addAriaLabel={ariaLabelAddTab}
+                    index={enabledTabsIndex + 1}
+                    onAddClick={this.wrapOnAddButton}
+                    tabIds={moreIds}
+                    isSelected={false}
+                    title={ariaLabelAddTab}
+                  />
+                  )}
+                </div>
               </div>
             )}
           </Droppable>
@@ -478,40 +637,53 @@ class Tabs extends React.Component {
     }
 
     return (
-      <div
-        {...attrs}
-        ref={this.containerRef}
-        className={commonTabsClassNames}
-        role="tablist"
-        aria-label={ariaLabel}
-        aria-orientation="horizontal"
-        aria-owns={hiddenIds.join(' ')}
-      >
-        {visibleTabs}
-        {this.showMoreButton ? (
-          <MoreButton
-            isOpen={this.isOpen}
-            hiddenIndex={this.hiddenStartIndex}
-            isActive={isHiddenSelected}
-            zIndex={this.state.visibleTabData.length - this.hiddenStartIndex}
-            onBlur={this.handleMoreButtonBlur}
-            onSelect={this.handleMoreButtonSelect}
-            refCallback={node => { this.moreButtonRef.current = node; }}
-            tabIds={ids}
-            variant={variant}
-          />
-        ) : undefined}
-        <TabDropDown
-          onFocus={this.handleHiddenFocus}
-          onBlur={this.handleHiddenBlur}
-          isOpen={this.isOpen}
-          onRequestClose={this.handleOutsideClick}
-          refCallback={node => { this.dropdownRef.current = node; }}
-          isDraggable={false}
-          onTabOrderChange={undefined}
+      <div className={commonTabsContainerClassNames}>
+        <div
+          {...attrs}
+          ref={this.containerRef}
+          className={commonTabsClassNames}
+          role="tablist"
+          aria-label={ariaLabel}
+          aria-orientation="horizontal"
+          aria-owns={hiddenIds.join(' ')}
         >
-          {hiddenTabs}
-        </TabDropDown>
+          {visibleTabs}
+          {this.showMoreButton ? (
+            <MoreButton
+              isOpen={this.isOpen}
+              hiddenIndex={this.hiddenStartIndex}
+              isActive={isHiddenSelected}
+              zIndex={this.state.visibleTabData.length - this.hiddenStartIndex}
+              onBlur={this.handleMoreButtonBlur}
+              onSelect={this.handleMoreButtonSelect}
+              refCallback={node => { this.moreButtonRef.current = node; }}
+              tabIds={ids}
+              variant={variant}
+            />
+          ) : undefined}
+          <TabDropDown
+            onFocus={this.handleHiddenFocus}
+            onBlur={this.handleHiddenBlur}
+            isOpen={this.isOpen}
+            onRequestClose={this.handleOutsideClick}
+            refCallback={node => { this.dropdownRef.current = node; }}
+          >
+            {hiddenTabs}
+          </TabDropDown>
+        </div>
+        <div className={commonDivClassNames} ref={this.addButtonRef}>
+          {(!this.showMoreButton && onSelectAddButton) && (
+          <AddButton
+            id={this.addTabId}
+            addAriaLabel={ariaLabelAddTab}
+            index={enabledTabsIndex + 1}
+            onAddClick={this.wrapOnAddButton}
+            tabIds={moreIds}
+            isSelected={false}
+            title={ariaLabelAddTab}
+          />
+          )}
+        </div>
       </div>
     );
   }