+++ packages/terra-tabs/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+## 7.6.0 - (September 5, 2023)
+
+* Added
+  * Added `isClosable` and `onTabClose` props to support close functionality in tabs.
+  * Added `onSelectAddButton` and `ariaLabelAddTab` props to support Add functionality in tabs.
+
 ## 7.5.0 - (August 31, 2023)
 
 * Added